# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/ticket":{"get":{"tags":["예매페이지"],"summary":"영화 리스트, 관람 연령가, 개봉날짜 예매율 순 출력","description":"예매 페이지 들어왔을때 첫 화면<b>정렬 : 예매율(투표율) 순 (기본)<br><hr><i>상영 예정작도 예매 가능하기 때문에 개봉 날짜가 필요할 것 같아서 추가로 넣었어여</hr></i>","operationId":"getAllMoviesForTickets","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieListForTickets"}}}}}}}},"/ticket/{movieId}":{"get":{"tags":["예매페이지"],"summary":"예매페이지에서 영화 선택시 아래 포스터 경로 출력","description":"예매페이지에서 영화 선택시 아래 포스터 경로 출력","operationId":"getMoviePoster","parameters":[{"name":"movieId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MoviePosterDTO"}}}}}}},"/ticket/abc":{"get":{"tags":["예매페이지"],"summary":"영화 리스트, 관람 연령가, 개봉날짜 예매율 순 출력","description":"예매 페이지 들어왔을때 첫 화면<b>정렬 : 가나다순 (선택)<br><hr><i>상영 예정작도 예매 가능하기 때문에 개봉 날짜가 필요할 것 같아서 추가로 넣었어여</hr></i>","operationId":"getAllMoviesForTicketsBydic","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieListForTickets"}}}}}}}},"/movies":{"get":{"tags":["메인페이지"],"summary":"영화 리스트 출력","description":"[상영 예정 영화]와 [현재 상영중인 영화] 다 합쳐서 인기순으로 정렬 <br> 해당 api에 들어올때마다 새로운 영화 데이터 받아서 다시 DB 업데이트 하는거라 시간이 좀 걸림<br> 'tmdb list page 1' 중 우리나라에서 [현재 방영중 + 상영 에정]인 영화만 골라서 저장돼서, 출력되는 영화 양이 적을 수 있음","operationId":"getAllMovies","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieDTOBuilder"}}}}}}}},"/movies/upcoming":{"get":{"tags":["메인페이지"],"summary":"상영 예정인 영화 리스트 출력","description":"[상영 예정인 영화] 인기순으로 정렬","operationId":"getUpcomingMovies","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieDTOBuilder"}}}}}}}},"/movies/nowplaying":{"get":{"tags":["메인페이지"],"summary":"현재 상영 중인 영화 리스트 출력","description":"[현재 상영중인 영화] 인기순으로 정렬","operationId":"getPlayingMovies","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieDTOBuilder"}}}}}}}},"/movies/details/{movieId}":{"get":{"tags":["메인페이지"],"summary":"영화 상세 정보 출력","description":"영화 상세 정보 출력<br><h3>조회 결과</h1><ul><li>\"movieId\" : 영화 ID</li><li>\"title\" : 영화 제목</li><li>\"posterPath\" : 포스터경로</li><li>\"releaseDate\" : 개봉 날짜</li><li>\"age\" : 관람연령</li><li>\"voteAverage\" : 투표율(예매율 대체)</li><li>\"productionCountry\" : 생산 국가</li><li>\"credits\" : 출연진(<i>\"director\" : 감독, \"character\": 캐릭터 - \"name\" : 배우</i>)</li><li>\"genres\" : 장르</li><li>\"overview\" : 소개 출력</li></ul>","operationId":"getMovieDetails","parameters":[{"name":"movieId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovieDetailsDTO"}}}}}}}},"components":{"schemas":{"MovieListForTickets":{"type":"object","properties":{"movieId":{"type":"integer","format":"int64"},"title":{"type":"string"},"age":{"type":"string"},"releaseDate":{"type":"string","format":"date"}}},"MoviePosterDTO":{"type":"object","properties":{"movieId":{"type":"integer","format":"int64"},"title":{"type":"string"},"posterPath":{"type":"string"},"releaseDate":{"type":"string","format":"date"}}},"MovieDTOBuilder":{"type":"object","properties":{"movie_id":{"type":"integer","format":"int64"},"poster_path":{"type":"string"},"title":{"type":"string"},"release_date":{"type":"string","format":"date"},"age":{"type":"string"},"vote_average":{"type":"number","format":"float"}}},"MovieDetailsDTO":{"type":"object","properties":{"movieId":{"type":"integer","format":"int64"},"posterPath":{"type":"string"},"title":{"type":"string"},"releaseDate":{"type":"string","format":"date"},"age":{"type":"string"},"voteAverage":{"type":"number","format":"float"},"productionCountry":{"type":"string"},"credits":{"type":"array","items":{"type":"object","additionalProperties":{"type":"string"}}},"genres":{"type":"array","items":{"type":"string"}},"overview":{"type":"string"},"runtime":{"type":"integer","format":"int32"}}}}}}